third_party/blink/web_tests/platform/mac-mac14/external/wpt/css/css-scrollbars/scrollbar-width-002-expected.txt

This is a testharness.js-based test.
[FAIL] vertical-lr, scrollbar-width auto
  assert_less_than: vertical-lr auto scrollHeight expected a number less than 200 but got 200
[FAIL] vertical-lr, scrollbar-width thin
  assert_less_than: vertical-lr thin scrollHeight expected a number less than 200 but got 200
[FAIL] vertical-rl, scrollbar-width auto
  assert_less_than: vertical-rl auto scrollHeight expected a number less than 200 but got 200
[FAIL] vertical-rl, scrollbar-width thin
  assert_less_than: vertical-rl thin scrollHeight expected a number less than 200 but got 200
Harness: the test ran to completion.
